Updated descriptor heap manager to allow for blocks of allocation ready for use with descriptor tables

+++ WaveE/Engine/WUploadManager.cpp
@@ -133,7 +133,7 @@ namespace WaveE
 		return newBefferIndex;
 	}
 
-	void WUploadManager::WaitForUpload(UINT bufferIndex)
+	void WUploadManager::WaitForUpload(UINT64 bufferIndex)
 	{
 		if (m_vUploadBuffers[bufferIndex].pFence->GetCompletedValue() < m_vUploadBuffers[bufferIndex].fenceValue)
 		{
